# service/swagger/ducklake-api.yaml
openapi: 3.0.3
info:
  title: DuckLake Data Lakehouse API
  description: |
    REST API for DuckLake - Git-like branching for data lakehouses.

    ## Overview
    This API provides endpoints to:
    - Execute SQL queries and return tabular results
    - Manage data branches (create, delete, search, switch)
    - Query branch metadata and statistics
    - Perform time-travel queries
    - Connect to multiple catalog types (DuckLake, Databricks, Snowflake)

    ## Catalog Types
    The service supports multiple catalog types:
    - **DUCKLAKE**: Native DuckLake catalogs with full read/write access and branching support
    - **DATABRICKS**: Databricks Unity Catalog (read-only, uses OAuth service principal)
    - **SNOWFLAKE**: Snowflake database (read-only, not yet implemented)

    ## Authentication
    All endpoints require Bearer token authentication.

    ## Error Handling
    Errors follow Google's gRPC status code conventions.

  version: 1.0.0
  contact:
    name: DuckLake Team
    email: support@ducklake.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ducklake.io/v1
    description: Production server
  - url: https://api.staging.ducklake.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /catalogs/{catalog_id}/branches:
    get:
      tags:
        - Branches
      summary: List branches
      description: |
        Returns a list of all branches in the catalog.

        **Note:** Branching is only supported for DUCKLAKE catalogs.
        For DATABRICKS and SNOWFLAKE catalogs, this endpoint will return an error.
      operationId: listBranches
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageNumber'
        - name: status
          in: query
          description: Filter by branch status
          schema:
            type: string
            enum: [active, merged, archived, deleted]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBranchesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Branches
      summary: Create branch
      description: |
        Creates a new branch from an existing branch at a specific snapshot.

        **Note:** Branching is only supported for DUCKLAKE catalogs.
      operationId: createBranch
      parameters:
        - $ref: '#/components/parameters/CatalogId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchRequest'
      responses:
        '201':
          description: Branch created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches:search:
    post:
      tags:
        - Branches
      summary: Search branches
      description: |
        Search branches with flexible filtering by name pattern, status,
        creation date, and parent branch.
      operationId: searchBranches
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchBranchesRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBranchesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches/{branch_name}:
    get:
      tags:
        - Branches
      summary: Get branch
      description: Returns details of a specific branch.
      operationId: getBranch
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/BranchName'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Branches
      summary: Delete branch
      description: Soft-deletes a branch (sets status to 'deleted').
      operationId: deleteBranch
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/BranchName'
      responses:
        '204':
          description: Branch deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches/{branch_name}:use:
    post:
      tags:
        - Branches
      summary: Switch to branch
      description: Sets the working branch for subsequent write operations.
      operationId: useBranch
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/BranchName'
      responses:
        '200':
          description: Branch switched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UseBranchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # BRANCH STATS & METADATA
  # ============================================================================
  /catalogs/{catalog_id}/branches:count:
    get:
      tags:
        - Branch Stats
      summary: Count branches
      description: Returns the total count of branches, optionally filtered by status.
      operationId: countBranches
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - name: status
          in: query
          description: Filter by branch status
          schema:
            type: string
            enum: [active, merged, archived, deleted]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchCountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches/{branch_name}/stats:
    get:
      tags:
        - Branch Stats
      summary: Get branch statistics
      description: |
        Returns detailed statistics for a branch including table count,
        schema count, file count, row count, and size.
      operationId: getBranchStats
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/BranchName'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches/{branch_name}/lineage:
    get:
      tags:
        - Branch Stats
      summary: Get branch lineage
      description: Returns the ancestry chain of a branch from root to current.
      operationId: getBranchLineage
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - $ref: '#/components/parameters/BranchName'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchLineageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches:activity:
    get:
      tags:
        - Branch Stats
      summary: List branches by activity
      description: Returns branches sorted by activity (last modified, created, or snapshot count).
      operationId: getBranchActivity
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - name: order_by
          in: query
          description: Field to order by
          schema:
            type: string
            enum: [last_modified, created, snapshot_count]
            default: last_modified
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchActivityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /catalogs/{catalog_id}/branches:by-age:
    get:
      tags:
        - Branch Stats
      summary: List branches by age
      description: |
        Returns branches filtered by age.
        Positive days = older than N days, negative = newer than N days.
      operationId: getBranchesByAge
      parameters:
        - $ref: '#/components/parameters/CatalogId'
        - name: days
          in: query
          required: true
          description: Number of days (positive=older than, negative=newer than)
          schema:
            type: integer
            example: -7
        - name: status
          in: query
          description: Filter by branch status
          schema:
            type: string
            enum: [active, merged, archived, deleted]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBranchesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  # --------------------------------------------------------------------------
  # PARAMETERS
  # --------------------------------------------------------------------------
  parameters:
    CatalogId:
      name: catalog_id
      in: path
      required: true
      description: Unique identifier for the DuckLake catalog
      schema:
        type: string
        example: "my_lakehouse"
    BranchName:
      name: branch_name
      in: path
      required: true
      description: Name of the branch
      schema:
        type: string
        example: "feature_analytics"
...
    PageSize:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PageNumber:
      name: page_number
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1

  # --------------------------------------------------------------------------
  # SCHEMAS
  # --------------------------------------------------------------------------
  schemas:
...
    ListBranchesResponse:
      type: object
      properties:
        branches:
          type: array
          items:
            $ref: '#/components/schemas/Branch'
        page:
          $ref: '#/components/schemas/Page'
...
    BranchCountResponse:
      type: object
      properties:
        count:
          type: integer
          format: int64
          description: Number of branches
          example: 5
        status_filter:
          type: string
          description: Status filter applied (or 'all')
          example: "active"
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication (gRPC code 16 - UNAUTHENTICATED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            unauthenticated:
              summary: No token provided
              value:
                code: 16
                message: "Authentication required"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "CREDENTIALS_MISSING"
                    domain: "ducklake.io"
            tokenExpired:
              summary: Token expired
              value:
                code: 16
                message: "Token has expired"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "TOKEN_EXPIRED"
                    domain: "ducklake.io"
...
    NotFound:
      description: Resource not found (gRPC code 5 - NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            catalogNotFound:
              summary: Catalog not found
              value:
                code: 5
                message: "Catalog not found"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Catalog"
                    resource_name: "nonexistent_catalog"
                    owner: "ducklake-service"
                    description: "The requested catalog does not exist"
            branchNotFound:
              summary: Branch not found
              value:
                code: 5
                message: "Branch not found"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Branch"
                    resource_name: "feature_nonexistent"
                    owner: "ducklake-service"
                    description: "The requested branch does not exist"
            tableNotFound:
              summary: Table not found
              value:
                code: 5
                message: "Table not found"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Table"
                    resource_name: "nonexistent_table"
                    owner: "ducklake-service"
                    description: "The requested table does not exist"
...
    InternalServerError:
      description: Internal server error (gRPC code 13 - INTERNAL)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            internalError:
              summary: Internal error
              value:
                code: 13
                message: "Internal server error"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "INTERNAL_ERROR"
                    domain: "ducklake.io"
            metadataConnectionError:
              summary: Metadata database connection error
              value:
                code: 14
                message: "Unable to connect to metadata database"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "METADATA_CONNECTION_ERROR"
                    domain: "ducklake.io"
                    metadata:
                      retry_after: "30"
            storageError:
              summary: Storage access error
              value:
                code: 13
                message: "Unable to access data storage"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "STORAGE_ERROR"
                    domain: "ducklake.io"